requirements.txt
click==8.1.7
coverage==7.6.4
exceptiongroup==1.2.2
iniconfig==2.0.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
Pygments==2.18.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-magic==0.4.27
PyYAML==6.0.2
rich==13.9.4
tomli==2.0.2
typing_extensions==4.12.2
